Now playing queue: don't close on click, and do nothing on long-click.
Doing nothing on long click allows to "cancel" a click if one puts their
finger on something and changes their mind: just hold on to long-click out of it.

Not closing (finish()) on click is useful for quickly scanning/browsing:
"was it this one?" ... "oh, wait, maybe this one" ... "no, let's play this other one instead".
It's annoying to have to re-open the view each time in this scenario, and since the
long click is not available anymore for this, and I personally don't see the point
anyway, we just keep the view on when we get a click.